fix: restructure export for not messing up build structure
