--- config/index-names.ts
@@ -1,7 +1,0 @@
-export const indexNames = {
-  USER_DOCS: 'sentry-docs-v2',
-  DEVELOP_DOCS: 'develop-docs',
-  HELP_CENTER: 'sentry-help',
-  ZENDESK: 'zendesk_sentry_articles',
-  BLOG: 'sentry-blog-posts',
-};
